+++ src/config/recommended.ts
@@ -7,6 +7,7 @@ module.exports = {
 		'deprecated': 'error',
 		'element-name': 'error',
 		'element-permitted-content': 'error',
+		'element-permitted-occurrences': 'error',
 		'no-dup-attr': 'error',
 		'no-dup-class': 'error',
 		'no-dup-id': 'error',